Extract `ReleaseNotesMetadataBlock` to a dedicated file
